Extract to separate ExpressionModel
