+++ plugins/ru.runa.gpd/src/ru/runa/gpd/ui/dialog/GlobalValidatorExpressionConstructorDialog.java
@@ -519,7 +519,7 @@ public class GlobalValidatorExpressionConstructorDialog extends Dialog {
     }
 
     private void getParamsMapFromExpression() throws UserFriendlyException {
-        Optional<GlobalValidatorExpressionConstructorDialog.ExpressionModel> optionalModel = GroovyCodeParser
+        Optional<ExpressionModel> optionalModel = GroovyCodeParser
                 .parseValidationModel(this.initializationExpression, this.variables);
         if (!optionalModel.isPresent()) {
             if (this.initializationExpression.equals("")) {
@@ -528,7 +528,7 @@ public class GlobalValidatorExpressionConstructorDialog extends Dialog {
             }
             throw this.exceptionFromGetParamsMapFromExpression();
         }
-        GlobalValidatorExpressionConstructorDialog.ExpressionModel model = optionalModel.get();
+        ExpressionModel model = optionalModel.get();
         this.initializationParams = model.getParamsMap();
     }
 
@@ -658,44 +658,5 @@ public class GlobalValidatorExpressionConstructorDialog extends Dialog {
     public void initializeFromExpression(String initializationExpression) {
         this.initializationExpression = initializationExpression;
         this.initializationType = InitializationType.EXPRESSION;
-    }
-
-    public static class ExpressionModel {
-        // модель выражения такого типа, который задается в этом диалоге.
-        // класс используется только для передачи данных о модели при загрузке из сохранения
-        private Map<String, String> params = null;
-
-        public ExpressionModel() {
-            this.params = new HashMap<>();
-            this.params.put(isSavedParam, "true");
-            this.params.put(expressionLineNumberParam, String.valueOf(0));
-        }
-
-        // возвращает представление модели в виде Map из параметров
-        public Map<String, String> getParamsMap() {
-            return params;
-        }
-
-        public int getExpressionLineNumber() {
-            return Integer.parseInt(this.params.get(expressionLineNumberParam));
-        }
-
-        public void addExpressionLineModel(ExpressionLine.ExpressionLineModel model) {
-            String expressionLinesNumberString = this.params.get(expressionLineNumberParam);
-            int expressionLinesNumber = Integer.parseInt(expressionLinesNumberString);
-            for (Entry<String, String> entry : model.getParamsMap().entrySet()) {
-                this.params.put(paramPrefix + "_" + expressionLinesNumberString + "_" + entry.getKey(), entry.getValue());
-            }
-            this.params.put(expressionLineNumberParam, String.valueOf(expressionLinesNumber + 1));
-        }
-
-        public ExpressionLine.ExpressionLineModel getExpressionLineModel(int index) {
-            ExpressionLine.ExpressionLineModel model = new ExpressionLine.ExpressionLineModel();
-            for (String param : ExpressionLine.getParamsNames()) {
-                String key = paramPrefix + "_" + index + "_" + param;
-                model.getParamsMap().put(param, this.params.get(key));
-            }
-            return model;
-        }
     }
 }
